updating MoveShape to alpha2
