+++ MoveShape/MoveShape/wwwroot/lib/signalr-clientES5-1.0.0-alpha2-final.js
@@ -95,43 +95,57 @@ var TextMessageFormat;
 })(TextMessageFormat = exports.TextMessageFormat || (exports.TextMessageFormat = {}));
 var BinaryMessageFormat;
 (function (BinaryMessageFormat) {
+    // The length prefix of binary messages is encoded as VarInt. Read the comment in
+    // the BinaryMessageParser.TryParseMessage for details.
     function write(output) {
-        // .byteLength does is undefined in IE10
+        // msgpack5 uses returns Buffer instead of Uint8Array on IE10 and some other browser
+        //  in which case .byteLength does will be undefined
         var size = output.byteLength || output.length;
-        var buffer = new Uint8Array(size + 8);
-        // javascript bitwise operators only support 32-bit integers
-        for (var i = 7; i >= 4; i--) {
-            buffer[i] = size & 0xff;
-            size = size >> 8;
-        }
-        buffer.set(output, 8);
+        var lenBuffer = [];
+        do {
+            var sizePart = size & 0x7f;
+            size = size >> 7;
+            if (size > 0) {
+                sizePart |= 0x80;
+            }
+            lenBuffer.push(sizePart);
+        } while (size > 0);
+        // msgpack5 uses returns Buffer instead of Uint8Array on IE10 and some other browser
+        //  in which case .byteLength does will be undefined
+        size = output.byteLength || output.length;
+        var buffer = new Uint8Array(lenBuffer.length + size);
+        buffer.set(lenBuffer, 0);
+        buffer.set(output, lenBuffer.length);
         return buffer.buffer;
     }
     BinaryMessageFormat.write = write;
     function parse(input) {
         var result = [];
         var uint8Array = new Uint8Array(input);
-        // 8 - the length prefix size
+        var maxLengthPrefixSize = 5;
+        var numBitsToShift = [0, 7, 14, 21, 28];
         for (var offset = 0; offset < input.byteLength;) {
-            if (input.byteLength < offset + 8) {
-                throw new Error("Cannot read message size");
-            }
-            // Note javascript bitwise operators only support 32-bit integers - for now cutting bigger messages.
-            // Tracking bug https://github.com/aspnet/SignalR/issues/613
-            if (!(uint8Array[offset] == 0 && uint8Array[offset + 1] == 0 && uint8Array[offset + 2] == 0 && uint8Array[offset + 3] == 0 && (uint8Array[offset + 4] & 0x80) == 0)) {
-                throw new Error("Messages bigger than 2147483647 bytes are not supported");
-            }
+            var numBytes = 0;
             var size = 0;
-            for (var i = 4; i < 8; i++) {
-                size = size << 8 | uint8Array[offset + i];
+            var byteRead = void 0;
+            do {
+                byteRead = uint8Array[offset + numBytes];
+                size = size | (byteRead & 0x7f) << numBitsToShift[numBytes];
+                numBytes++;
+            } while (numBytes < Math.min(maxLengthPrefixSize, input.byteLength - offset) && (byteRead & 0x80) != 0);
+            if ((byteRead & 0x80) !== 0 && numBytes < maxLengthPrefixSize) {
+                throw new Error("Cannot read message size.");
             }
-            if (uint8Array.byteLength >= offset + 8 + size) {
+            if (numBytes === maxLengthPrefixSize && byteRead > 7) {
+                throw new Error("Messages bigger than 2GB are not supported.");
+            }
+            if (uint8Array.byteLength >= offset + numBytes + size) {
                 // IE does not support .slice() so use subarray
-                result.push(uint8Array.slice ? uint8Array.slice(offset + 8, offset + 8 + size) : uint8Array.subarray(offset + 8, offset + 8 + size));
+                result.push(uint8Array.slice ? uint8Array.slice(offset + numBytes, offset + numBytes + size) : uint8Array.subarray(offset + numBytes, offset + numBytes + size));
             } else {
-                throw new Error("Incomplete message");
+                throw new Error("Incomplete message.");
             }
-            offset = offset + 8 + size;
+            offset = offset + numBytes + size;
         }
         return result;
     }
@@ -342,8 +356,8 @@ var HttpConnection = function () {
                             case 8:
                                 this.url += (this.url.indexOf("?") == -1 ? "?" : "&") + ("id=" + this.connectionId);
                                 this.transport = this.createTransport(this.options.transport, negotiateResponse.availableTransports);
-                                this.transport.onDataReceived = this.onDataReceived;
-                                this.transport.onClosed = function (e) {
+                                this.transport.onreceive = this.onreceive;
+                                this.transport.onclose = function (e) {
                                     return _this.stopConnection(true, e);
                                 };
                                 requestedTransferMode = this.features.transferMode === 2 /* Binary */
@@ -466,8 +480,8 @@ var HttpConnection = function () {
                 this.transport = null;
             }
             this.connectionState = 3 /* Disconnected */;
-            if (raiseClosed && this.onClosed) {
-                this.onClosed(error);
+            if (raiseClosed && this.onclose) {
+                this.onclose(error);
             }
         }
     }, {
@@ -477,12 +491,15 @@ var HttpConnection = function () {
             if (url.lastIndexOf("https://", 0) === 0 || url.lastIndexOf("http://", 0) === 0) {
                 return url;
             }
-            if (typeof window === 'undefined') {
+            if (typeof window === 'undefined' || !window || !window.document) {
                 throw new Error("Cannot resolve '" + url + "'.");
             }
             var parser = window.document.createElement("a");
             parser.href = url;
             var baseUrl = !parser.protocol || parser.protocol === ":" ? window.document.location.protocol + "//" + (parser.host || window.document.location.host) : parser.protocol + "//" + parser.host;
+            if (!url || url[0] != '/') {
+                url = '/' + url;
+            }
             var normalizedUrl = baseUrl + url;
             this.logger.log(ILogger_1.LogLevel.Information, "Normalizing '" + url + "' to '" + normalizedUrl + "'");
             return normalizedUrl;
@@ -625,20 +642,21 @@ var HubConnection = function () {
         }
         this.logger = Loggers_1.LoggerFactory.createLogger(options.logging);
         this.protocol = options.protocol || new JsonHubProtocol_1.JsonHubProtocol();
-        this.connection.onDataReceived = function (data) {
-            _this.onDataReceived(data);
+        this.connection.onreceive = function (data) {
+            return _this.processIncomingData(data);
         };
-        this.connection.onClosed = function (error) {
-            _this.onConnectionClosed(error);
+        this.connection.onclose = function (error) {
+            return _this.connectionClosed(error);
         };
         this.callbacks = new _map2.default();
         this.methods = new _map2.default();
+        this.closedCallbacks = [];
         this.id = 0;
     }
 
     (0, _createClass3.default)(HubConnection, [{
-        key: "onDataReceived",
-        value: function onDataReceived(data) {
+        key: "processIncomingData",
+        value: function processIncomingData(data) {
             // Parse the messages
             var messages = this.protocol.parseMessages(data);
             for (var i = 0; i < messages.length; ++i) {
@@ -666,9 +684,13 @@ var HubConnection = function () {
     }, {
         key: "invokeClientMethod",
         value: function invokeClientMethod(invocationMessage) {
-            var method = this.methods.get(invocationMessage.target.toLowerCase());
-            if (method) {
-                method.apply(this, invocationMessage.arguments);
+            var _this2 = this;
+
+            var methods = this.methods.get(invocationMessage.target.toLowerCase());
+            if (methods) {
+                methods.forEach(function (m) {
+                    return m.apply(_this2, invocationMessage.arguments);
+                });
                 if (!invocationMessage.nonblocking) {
                     // TODO: send result back to the server?
                 }
@@ -677,8 +699,10 @@ var HubConnection = function () {
             }
         }
     }, {
-        key: "onConnectionClosed",
-        value: function onConnectionClosed(error) {
+        key: "connectionClosed",
+        value: function connectionClosed(error) {
+            var _this3 = this;
+
             var errorCompletionMessage = {
                 type: 3 /* Completion */
                 , invocationId: "-1",
@@ -688,9 +712,9 @@ var HubConnection = function () {
                 callback(errorCompletionMessage);
             });
             this.callbacks.clear();
-            if (this.connectionClosedCallback) {
-                this.connectionClosedCallback(error);
-            }
+            this.closedCallbacks.forEach(function (c) {
+                return c.apply(_this3, [error]);
+            });
         }
     }, {
         key: "start",
@@ -735,7 +759,7 @@ var HubConnection = function () {
     }, {
         key: "stream",
         value: function stream(methodName) {
-            var _this2 = this;
+            var _this4 = this;
 
             for (var _len = arguments.length, args = Array(_len > 1 ? _len - 1 : 0), _key = 1; _key < _len; _key++) {
                 args[_key - 1] = arguments[_key];
@@ -761,7 +785,7 @@ var HubConnection = function () {
             var message = this.protocol.writeMessage(invocationDescriptor);
             this.connection.send(message).catch(function (e) {
                 subject.error(e);
-                _this2.callbacks.delete(invocationDescriptor.invocationId);
+                _this4.callbacks.delete(invocationDescriptor.invocationId);
             });
             return subject;
         }
@@ -779,7 +803,7 @@ var HubConnection = function () {
     }, {
         key: "invoke",
         value: function invoke(methodName) {
-            var _this3 = this;
+            var _this5 = this;
 
             for (var _len3 = arguments.length, args = Array(_len3 > 1 ? _len3 - 1 : 0), _key3 = 1; _key3 < _len3; _key3++) {
                 args[_key3 - 1] = arguments[_key3];
@@ -787,7 +811,7 @@ var HubConnection = function () {
 
             var invocationDescriptor = this.createInvocation(methodName, args, false);
             var p = new _promise2.default(function (resolve, reject) {
-                _this3.callbacks.set(invocationDescriptor.invocationId, function (invocationEvent) {
+                _this5.callbacks.set(invocationDescriptor.invocationId, function (invocationEvent) {
                     if (invocationEvent.type === 3 /* Completion */) {
                             var completionMessage = invocationEvent;
                             if (completionMessage.error) {
@@ -799,10 +823,10 @@ var HubConnection = function () {
                         reject(new Error("Streaming methods must be invoked using HubConnection.stream"));
                     }
                 });
-                var message = _this3.protocol.writeMessage(invocationDescriptor);
-                _this3.connection.send(message).catch(function (e) {
+                var message = _this5.protocol.writeMessage(invocationDescriptor);
+                _this5.connection.send(message).catch(function (e) {
                     reject(e);
-                    _this3.callbacks.delete(invocationDescriptor.invocationId);
+                    _this5.callbacks.delete(invocationDescriptor.invocationId);
                 });
             });
             return p;
@@ -810,7 +834,37 @@ var HubConnection = function () {
     }, {
         key: "on",
         value: function on(methodName, method) {
-            this.methods.set(methodName.toLowerCase(), method);
+            if (!methodName || !method) {
+                return;
+            }
+            methodName = methodName.toLowerCase();
+            if (!this.methods.has(methodName)) {
+                this.methods.set(methodName, []);
+            }
+            this.methods.get(methodName).push(method);
+        }
+    }, {
+        key: "off",
+        value: function off(methodName, method) {
+            if (!methodName || !method) {
+                return;
+            }
+            methodName = methodName.toLowerCase();
+            var handlers = this.methods.get(methodName);
+            if (!handlers) {
+                return;
+            }
+            var removeIdx = handlers.indexOf(method);
+            if (removeIdx != -1) {
+                handlers.splice(removeIdx, 1);
+            }
+        }
+    }, {
+        key: "onclose",
+        value: function onclose(callback) {
+            if (callback) {
+                this.closedCallbacks.push(callback);
+            }
         }
     }, {
         key: "createInvocation",
@@ -824,11 +878,6 @@ var HubConnection = function () {
                 arguments: args,
                 nonblocking: nonblocking
             };
-        }
-    }, {
-        key: "onClosed",
-        set: function set(callback) {
-            this.connectionClosedCallback = callback;
         }
     }]);
     return HubConnection;
@@ -1186,17 +1235,17 @@ var WebSocketTransport = function () {
                 };
                 webSocket.onmessage = function (message) {
                     _this.logger.log(ILogger_1.LogLevel.Trace, "(WebSockets transport) data received: " + message.data);
-                    if (_this.onDataReceived) {
-                        _this.onDataReceived(message.data);
+                    if (_this.onreceive) {
+                        _this.onreceive(message.data);
                     }
                 };
                 webSocket.onclose = function (event) {
                     // webSocket will be null if the transport did not start successfully
-                    if (_this.onClosed && _this.webSocket) {
+                    if (_this.onclose && _this.webSocket) {
                         if (event.wasClean === false || event.code !== 1000) {
-                            _this.onClosed(new Error("Websocket closed with status code: " + event.code + " (" + event.reason + ")"));
+                            _this.onclose(new Error("Websocket closed with status code: " + event.code + " (" + event.reason + ")"));
                         } else {
-                            _this.onClosed();
+                            _this.onclose();
                         }
                     }
                 };
@@ -1246,13 +1295,13 @@ var ServerSentEventsTransport = function () {
                 var eventSource = new EventSource(_this2.url);
                 try {
                     eventSource.onmessage = function (e) {
-                        if (_this2.onDataReceived) {
+                        if (_this2.onreceive) {
                             try {
                                 _this2.logger.log(ILogger_1.LogLevel.Trace, "(SSE transport) data received: " + e.data);
-                                _this2.onDataReceived(e.data);
+                                _this2.onreceive(e.data);
                             } catch (error) {
-                                if (_this2.onClosed) {
-                                    _this2.onClosed(error);
+                                if (_this2.onclose) {
+                                    _this2.onclose(error);
                                 }
                                 return;
                             }
@@ -1261,8 +1310,8 @@ var ServerSentEventsTransport = function () {
                     eventSource.onerror = function (e) {
                         reject();
                         // don't report an error if the transport did not start successfully
-                        if (_this2.eventSource && _this2.onClosed) {
-                            _this2.onClosed(new Error(e.message || "Error occurred"));
+                        if (_this2.eventSource && _this2.onclose) {
+                            _this2.onclose(new Error(e.message || "Error occurred"));
                         }
                     };
                     eventSource.onopen = function () {
@@ -1339,38 +1388,38 @@ var LongPollingTransport = function () {
             var pollXhr = new XMLHttpRequest();
             pollXhr.onload = function () {
                 if (pollXhr.status == 200) {
-                    if (_this3.onDataReceived) {
+                    if (_this3.onreceive) {
                         try {
                             var response = transferMode === 1 /* Text */
                             ? pollXhr.responseText : pollXhr.response;
                             if (response) {
                                 _this3.logger.log(ILogger_1.LogLevel.Trace, "(LongPolling transport) data received: " + response);
-                                _this3.onDataReceived(response);
+                                _this3.onreceive(response);
                             } else {
                                 _this3.logger.log(ILogger_1.LogLevel.Information, "(LongPolling transport) timed out");
                             }
                         } catch (error) {
-                            if (_this3.onClosed) {
-                                _this3.onClosed(error);
+                            if (_this3.onclose) {
+                                _this3.onclose(error);
                             }
                             return;
                         }
                     }
                     _this3.poll(url, transferMode);
                 } else if (_this3.pollXhr.status == 204) {
-                    if (_this3.onClosed) {
-                        _this3.onClosed();
+                    if (_this3.onclose) {
+                        _this3.onclose();
                     }
                 } else {
-                    if (_this3.onClosed) {
-                        _this3.onClosed(new HttpError_1.HttpError(pollXhr.statusText, pollXhr.status));
+                    if (_this3.onclose) {
+                        _this3.onclose(new HttpError_1.HttpError(pollXhr.statusText, pollXhr.status));
                     }
                 }
             };
             pollXhr.onerror = function () {
-                if (_this3.onClosed) {
+                if (_this3.onclose) {
                     // network related error or denied cross domain request
-                    _this3.onClosed(new Error("Sending HTTP request failed."));
+                    _this3.onclose(new Error("Sending HTTP request failed."));
                 }
             };
             pollXhr.ontimeout = function () {
